fix tests for r5559 on databases not supporting check constraints
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@5561 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ dsmf/src/com/exedio/dsmf/Table.java
@@ -331,7 +331,7 @@ public final class Table extends Node
 	{
 		for(final Constraint constraint : constraintList)
 		{
-			if(constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
+			if(constraint.isSupported() && constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
 				constraint.create();
 		}
 	}
@@ -340,7 +340,7 @@ public final class Table extends Node
 	{
 		for(final Constraint constraint : constraintList)
 		{
-			if(constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
+			if(constraint.isSupported() && constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
 				constraint.drop();
 		}
 	}
